# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "使用gin+vue进行极速开发的全栈开发基础平台",
    "title": "Gin-Vue-Admin Swagger API接口文档",
    "contact": {},
    "version": "v2.5.9"
  },
  "paths": {
...
    "/autoCode/createPackage": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "创建package",
        "parameters": [
          {
            "description": "创建package",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAutoCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建package成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/autoCode/createPlug": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "创建插件模板",
        "parameters": [
          {
            "description": "创建插件模板",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAutoCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建插件模板成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/autoCode/createTemp": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "自动代码模板",
        "parameters": [
          {
            "description": "创建自动代码",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.AutoCodeStruct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"创建成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/autoCode/delPackage": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "删除package",
        "parameters": [
          {
            "description": "创建package",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAutoCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除package成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/autoCode/getColumn": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "获取当前表所有字段",
        "responses": {
          "200": {
            "description": "获取当前表所有字段",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/autoCode/getDatabase": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "获取当前所有数据库",
        "responses": {
          "200": {
            "description": "获取当前所有数据库",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/autoCode/getPackage": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "获取package",
        "responses": {
          "200": {
            "description": "创建package成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
    "/autoCode/getTables": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "获取当前数据库所有表",
        "responses": {
          "200": {
            "description": "获取当前数据库所有表",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/autoCode/installPlugin": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "安装插件",
        "parameters": [
          {
            "type": "file",
            "description": "this is a test file",
            "name": "plug",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "安装插件成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/autoCode/preview": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "预览创建后的代码",
        "parameters": [
          {
            "description": "预览创建代码",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.AutoCodeStruct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "预览创建后的代码",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/autoCode/pubPlug": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["AutoCode"],
        "summary": "打包插件",
        "parameters": [
          {
            "description": "打包插件",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAutoCode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "打包插件成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "msg": {
          "type": "string"
        }
      }
    },
...
    "system.AutoCodeStruct": {
      "type": "object",
      "properties": {
        "abbreviation": {
          "description": "Struct简称",
          "type": "string"
        },
        "autoCreateApiToSql": {
          "description": "是否自动创建api",
          "type": "boolean"
        },
        "autoCreateResource": {
          "description": "是否自动创建资源标识",
          "type": "boolean"
        },
        "autoMoveFile": {
          "description": "是否自动移动文件",
          "type": "boolean"
        },
        "businessDB": {
          "description": "业务数据库",
          "type": "string"
        },
        "description": {
          "description": "Struct中文名称",
          "type": "string"
        },
        "fields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.Field"
          }
        },
        "hasTimer": {
          "type": "boolean"
        },
        "humpPackageName": {
          "description": "go文件名称",
          "type": "string"
        },
        "package": {
          "type": "string"
        },
        "packageName": {
          "description": "文件名称",
          "type": "string"
        },
        "structName": {
          "description": "Struct名称",
          "type": "string"
        },
        "tableName": {
          "description": "表名",
          "type": "string"
        }
      }
    },
    "system.Field": {
      "type": "object",
      "properties": {
        "clearable": {
          "description": "是否可清空",
          "type": "boolean"
        },
        "columnName": {
          "description": "数据库字段",
          "type": "string"
        },
        "comment": {
          "description": "数据库字段描述",
          "type": "string"
        },
        "dataTypeLong": {
          "description": "数据库字段长度",
          "type": "string"
        },
        "dictType": {
          "description": "字典",
          "type": "string"
        },
        "errorText": {
          "description": "校验失败文字",
          "type": "string"
        },
        "fieldDesc": {
          "description": "中文名",
          "type": "string"
        },
        "fieldJson": {
          "description": "FieldJson",
          "type": "string"
        },
        "fieldName": {
          "description": "Field名",
          "type": "string"
        },
        "fieldSearchType": {
          "description": "搜索条件",
          "type": "string"
        },
        "fieldType": {
          "description": "Field数据类型",
          "type": "string"
        },
        "require": {
          "description": "是否必填",
          "type": "boolean"
        },
        "sort": {
          "description": "是否增加排序",
          "type": "boolean"
        }
      }
    },
...
    "system.SysAutoCode": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "主键ID",
          "type": "integer"
        },
        "createdAt": {
          "description": "创建时间",
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "label": {
          "type": "string"
        },
        "packageName": {
          "type": "string"
        },
        "updatedAt": {
          "description": "更新时间",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "x-token",
      "in": "header"
    }
  }
}